changed double to float

+++ AggroSystem/Entity.cs
@@ -2,13 +2,13 @@ namespace AggroSystem
 {
     public interface Entity
     {
-        double DistanceSquared(Entity entity)
+        float DistanceSquared(Entity entity)
         {
             return (entity.GetXPos() - GetXPos()) * (entity.GetXPos() - GetXPos()) + 
             (entity.GetYPos() - GetYPos()) * (entity.GetYPos() - GetYPos());
         }
 
-        double GetXPos();
-        double GetYPos();
+        float GetXPos();
+        float GetYPos();
     }
 }
