+++ AggroSystem/SoundAggroable.cs
@@ -2,6 +2,6 @@ namespace AggroSystem
 {
     public interface SoundAggroable : Aggroable
     {
-        double GetMinimumSoundIntensity();
+        float GetMinimumSoundIntensity();
     }
 }
